refactor react
