+++ packages/common/src/index.ts
@@ -1,2 +1,2 @@
 export * from "./types";
-export * from "./messages";
+export * from "./events";
