--- packages/common/src/messages.ts
@@ -1,8 +1,0 @@
-
-export class ProviderSelectedEvent extends CustomEvent<{url: string}> {
-  constructor(providerUrl: string) {
-    super('providerSelected', {
-      detail: {url: providerUrl},
-    });
-  }
-}
