+++ packages/react/package.json
@@ -18,6 +18,7 @@
         "test:unit": "exit 0"
     },
     "dependencies": {
+        "@chainsafe/eth-sso-common": "workspace:^",
         "@chainsafe/eth-sso-ui": "workspace:^",
         "caip": "^1.1.1"
     },
